Handle via header and refactor code
Via header is a mandatory field to be added.
If the user does not want to expose their identity as a proxy server, user can provide the pseudonym that needs to be added to the via header.
If this pseudonym is not given, it will add the proxy server host to the via header.